Introduce distance_of_time_in_words.
This function is taken from Rails core and it's used to print
in a fancy way a distance of time expressed in seconds using
human readable words. As an example it has been used in frame
dissector to print the distance between the current packet and
the reference.

The need of round() function requires the libm to be added to
wsutil.

Change-Id: I06047b8f190f0cdea58fb54cbce7d2d3bef9c9f5